Timings
Started a new file to start linking video's to behavioral files.

2_preprocess has been changed to also include syncing beeps, and per trial start, stop and answer timings. New allPMSDATA.csv has been exported. Still to check whether RT complies with the change in start and stop stuff.

---------------------

Now a lot of code has been written but not yet tested because waiting on getFileDurations
The following: timings of starts and stops with correction for beep have been made for behavioral and video's. Then indices are made for the video's, and mean values are calculated for relevant features. These are added to allData variable.