BIG UPDATE
Okay some updates with regards to data per RQ, because the trait questionnaires don;t need to be excluded for the participants that didnt participate in the state stuff.
That's what is done here. New csv written as cleanedData.csv consisting of all relevant (AND COMPLETE) data for both the trait as the questionnaires at states.